Negative deadline values as relative timeouts
Treat a negative number deadline (other than -Infinity) as a millisecond timeout from the point the call is made. This makes it easier for client code to specify a relative timeout. It also allows a CallOptions object containing a relative timeout to be used unmodified for multiple gRPC calls taking place at different times, which would not be possible with a fixed time deadline.